requirements.txt
# tbt
torch
setuptools
python-dotenv